requirements.txt
glfw==2.5.3
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.6
